apps/api_py/pyproject.toml
[project]
name = "api-key-organizer-api-py"
version = "0.1.0"
description = "FastAPI backend for API Key Organizer"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.13.2",
  "cryptography>=43.0.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "passlib[argon2]>=1.7.4",
  "pydantic-settings>=2.5.2",
  "python-jose[cryptography]>=3.3.0",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "uvicorn[standard]>=0.30.6"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2",
  "pytest>=8.3.3"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
